Lesson code finished
Main lessons code is finished. For the other less it is as simple as changing variable names. Although there is an unexpected error when pressing the right arrow button multiple times than pressing the previous button, the labels over lap. But the labels are just mere placeholders at this stage